+++ rusvid_server/src/redis.rs
@@ -3,6 +3,6 @@ pub const fn video_status_prefix() -> &'static str {
     "video_id_"
 }
 
-pub fn key_for_video_status(video_id: &String) -> String {
-    format!("{}{video_id}", video_status_prefix())
+pub fn key_for_video_status(video_id: &str) -> String {
+    format!("{prefix}{video_id}", prefix = video_status_prefix())
 }